debut algo
Correction des room et création de la class de base pour la gestion de la map avec les listes nécessaire.
