+++ Assets/Scenes/Prototypes/proto_move_client.unity
@@ -121,6 +121,77 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
+--- !u!1 &624650637
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 624650639}
+  - component: {fileID: 624650638}
+  - component: {fileID: 624650640}
+  m_Layer: 0
+  m_Name: MasterGrid
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!156049354 &624650638
+Grid:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 624650637}
+  m_Enabled: 1
+  m_CellSize: {x: 1, y: 1, z: 0}
+  m_CellGap: {x: 0, y: 0, z: 0}
+  m_CellLayout: 0
+  m_CellSwizzle: 0
+--- !u!4 &624650639
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 624650637}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 5
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &624650640
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 624650637}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d48f061af1d36404ca1a03eb4181349d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_bondaries: {x: 100, y: 100}
+  m_numberRandomRoom: 5
+  m_roomPoolRandom:
+  - {fileID: 1808768653936752115, guid: 3d468dc9c994ab64c8cbc88f0779d4a1, type: 3}
+  - {fileID: 2217575440485873542, guid: 2146b954d84975444985edc8f23e8c79, type: 3}
+  - {fileID: 8628403321377010401, guid: 0a58398403efe9544ac88c0ea5212b99, type: 3}
+  - {fileID: 648168483729528984, guid: dc8b70697a5098b4cae0dda1f70658df, type: 3}
+  m_roomPoolSection0:
+  - {fileID: 805576736439278860, guid: 53d236ef3cac8d94f980759f252f34cc, type: 3}
+  - {fileID: 7375235168778441758, guid: dcf4c13dbfd5f2c4ab30f2fd1dd08476, type: 3}
+  - {fileID: 4131869434115816236, guid: 3928f1980e463394ea90c9a134d63319, type: 3}
+  - {fileID: 4131869434115816236, guid: 3928f1980e463394ea90c9a134d63319, type: 3}
+  - {fileID: 4131869434115816236, guid: 3928f1980e463394ea90c9a134d63319, type: 3}
+  - {fileID: 4131869434115816236, guid: 3928f1980e463394ea90c9a134d63319, type: 3}
+  m_roomPoolSection1: []
 --- !u!1 &925986282
 GameObject:
   m_ObjectHideFlags: 0
@@ -720,72 +791,3 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   m_playerTransform: {fileID: 1146798267}
---- !u!1001 &8628403321826270060
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 8628403321377010401, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_Name
-      value: Grid
-      objectReference: {fileID: 0}
-    - target: {fileID: 8628403321377010403, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8628403321377010403, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8628403321377010403, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8628403321377010403, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8628403321377010403, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8628403321377010403, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8628403321377010403, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 8628403321377010403, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 5
-      objectReference: {fileID: 0}
-    - target: {fileID: 8628403321377010403, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8628403321377010403, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8628403321377010403, guid: 0a58398403efe9544ac88c0ea5212b99,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 0a58398403efe9544ac88c0ea5212b99, type: 3}
