+++ Assets/Scripts/Common/World.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7da1d72b42bd10d4f89657b0d3d60b19
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
