+++ Assets/Scripts/Common/World/RoomInfo.cs
@@ -1,0 +1,21 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace ubv.common.world
+{
+    public class RoomInfo : MonoBehaviour
+    {
+
+        [SerializeField] private int m_height;
+        [SerializeField] private int m_width;
+
+        Vector2Int GetDimension()
+        {
+            return new Vector2Int(m_width, m_height);
+        }
+
+    }
+
+}
+
