+++ Assets/Scripts/Common/World/RoomInfo.cs.meta
