+++ Assets/Scripts/Common/World/WorldGenerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d48f061af1d36404ca1a03eb4181349d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
